Add a new algorithm for finding actual symbols
Old algorithm is still available in 'verify' mode, which can catch when
we resolve two different symbols to a signle function. In the new
algorithm, when a match if found, it is removed from the match set so we
cannot catch if any other symbols resolve to that function.

The new algorithm can be slightly faster, and also enables us to cache
symbol matching results.